You know what, this works. I'll fix the issues in the next PR. Promise!

+++ src/main/java/io/github/incplusplus/betterstat/service/ThermostatServiceImpl.java
@@ -3,6 +3,7 @@ package io.github.incplusplus.betterstat.service;
 import io.github.incplusplus.betterstat.persistence.model.Thermostat;
 import io.github.incplusplus.betterstat.persistence.repositories.ThermostatRepository;
 import io.github.incplusplus.betterstat.statemachine.StateHandler;
+import io.github.incplusplus.betterstat.web.exception.ObjectNotFoundException;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -29,6 +30,17 @@ public class ThermostatServiceImpl implements ThermostatService {
   }
 
   @Override
+  public Thermostat updateThermostat(String id, Thermostat thermostat)
+      throws ObjectNotFoundException {
+    if (!this.existsById(id)) throw new ObjectNotFoundException(id, Thermostat.class);
+    // If it already exists, set the ID of the DTO to the ID we know exists in our repository.
+    thermostat.setId(id);
+    // Saving an object with an ID that already exists will simply update it.
+    // https://stackoverflow.com/a/56207430/1687436
+    return thermostatRepository.save(thermostat);
+  }
+
+  @Override
   public Optional<Thermostat> getThermostatById(String id) {
     return thermostatRepository.findById(id);
   }
@@ -39,7 +51,17 @@ public class ThermostatServiceImpl implements ThermostatService {
   }
 
   @Override
-  public void deleteById(String id) {
+  public Thermostat deleteById(String id) throws ObjectNotFoundException {
+    Optional<Thermostat> thermostatOptional = this.getThermostatById(id);
+    if (thermostatOptional.isEmpty()) {
+      throw new ObjectNotFoundException(id, Thermostat.class);
+    }
     thermostatRepository.deleteById(id);
+    return thermostatOptional.get();
+  }
+
+  @Override
+  public boolean existsById(String id) {
+    return thermostatRepository.existsById(id);
   }
 }
